DOC-2608: Move rssfeed.cjs to lib folder and add error handling to prevent build from failing
